+++ packages/editor-common/web/src/Base/createBasePlugin.jsx
@@ -36,7 +36,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
     onOverlayClick,
     onAtomicBlockFocus,
     disableRightClick,
+    commonPubsub,
+    defaultPluginData,
+    pluginDefaults,
   } = config;
+  defaultPluginData && (pluginDefaults[config.type] = defaultPluginData);
   const toolbarTheme = { ...getToolbarTheme(config.theme, 'plugin'), ...config.theme };
   const Toolbar =
     config?.toolbar?.InlineButtons &&
@@ -67,9 +71,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
         button,
         helpers,
         pubsub,
+        commonPubsub,
         settings,
         t,
         isMobile,
+        pluginDefaults,
       }),
     }));
   const PluginComponent = config.component;
@@ -85,6 +91,7 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       onOverlayClick,
       onAtomicBlockFocus,
       pubsub,
+      commonPubsub,
       settings,
       helpers,
       t,