+++ packages/plugin-button/web/src/createButtonPlugin.js
@@ -2,7 +2,7 @@ import createToolbar from './toolbar';
 import { mergeStyles } from 'wix-rich-content-common';
 import { createBasePlugin } from 'wix-rich-content-editor-common';
 
-import { BUTTON_TYPE } from './constants';
+import { BUTTON_TYPE, DEFAULTS } from './constants';
 
 import Styles from '../statics/styles/default-styles.scss';
 import ButtonComponent from './components/button-component';
@@ -19,7 +19,9 @@ const createButtonPlugin = (config = {}) => {
     [type]: settings = {},
     ...rest
   } = config;
+
   const styles = mergeStyles({ styles: Styles, theme });
+
   return createBasePlugin({
     component: ButtonComponent,
     settings,
@@ -39,6 +41,7 @@ const createButtonPlugin = (config = {}) => {
     }),
     helpers,
     t,
+    defaultPluginData: DEFAULTS,
     ...rest,
   });
 };