+++ packages/plugin-file-upload/web/src/toolbar/insert-file-upload-buttons.jsx
@@ -7,6 +7,7 @@ export default ({ helpers, settings, t }) => {
   return [
     {
       type: 'file',
+      multi: true,
       name: 'UploadFile',
       tooltipText: t('FileUploadInsertButton_tooltip'),
       Icon: icon,