+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { isEqual } from 'lodash';
 import { mergeStyles, Context, validate } from 'wix-rich-content-common';
 import { convertItemData } from './helpers/convert-item-data';
-import { getDefault, isHorizontalLayout, sampleItems } from './constants';
+import { DEFAULTS, isHorizontalLayout, sampleItems } from './constants';
 import resizeMediaUrl from './helpers/resize-media-url';
 import schema from '../statics/data-schema.json';
 import styles from '../statics/styles/viewer.scss';
@@ -72,10 +72,9 @@ class GalleryViewer extends React.Component {
   };
 
   stateFromProps = props => {
-    const defaults = getDefault();
-    const items = props.componentData.items || defaults.items;
+    const items = props.componentData.items || DEFAULTS.items;
     const styleParams = this.getStyleParams(
-      Object.assign(defaults.styles, props.componentData.styles || {}),
+      Object.assign(DEFAULTS.styles, props.componentData.styles || {}),
       items
     );
     return {