+++ packages/plugin-video/web/src/createVideoPlugin.js
@@ -1,5 +1,5 @@
 import createToolbar from './toolbar';
-import { Component } from './video-component';
+import { Component, DEFAULTS } from './video-component';
 import { VIDEO_TYPE, VIDEO_TYPE_LEGACY } from './types';
 import { createBasePlugin } from 'wix-rich-content-editor-common';
 
@@ -22,6 +22,7 @@ const createVideoPlugin = (config = {}) => {
     t,
     isMobile,
     disableRightClick: config?.uiSettings?.disableRightClick,
+    defaultPluginData: DEFAULTS,
     ...rest,
   });
 };